% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangle.R
\name{setreuler_coordinates}
\alias{setreuler_coordinates}
\title{Generate coordinates for a Euler diagram}
\usage{
setreuler_coordinates(in_a, in_b, width = 10, offset = 0)
}
\arguments{
\item{in_a}{Logical vector for inclusion in set A.}

\item{in_b}{logical vector for inclusion in set B.}

\item{width}{width of the rectangle for set A.}

\item{offset}{horizontal offset for the rectangle of set B.}
}
\value{
A data frame of coordinates for a Euler diagram.
}
\description{
Generate coordinates for a Euler diagram
}
