% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangle.R
\name{setreuler_plot}
\alias{setreuler_plot}
\title{Plot coordinate data from setreuler}
\usage{
setreuler_plot(coordinates_data, radius = NULL)
}
\arguments{
\item{coordinates_data}{Data frame generated by
\code{setreuler_coordinates()}.}

\item{radius}{Radius for rounded corners on rectangles.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot coordinate data from setreuler
}
