% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangle.R
\name{venn_square}
\alias{venn_square}
\title{Venn diagram with squares}
\usage{
venn_square(data)
}
\arguments{
\item{data}{Data for diagram, passed to `venndata`}
}
\value{
Coordinates for the squares.
}
\description{
Venn diagram with squares
}
